LPD-26928 Merge nodes tests into a single file
